added ability to change dtypes

+++ src/montecarlo.py
@@ -15,7 +15,7 @@ def mc_propagate(
     Repeatedly propagate an initial field given by u_init_func through a propagation function prop_func.
     Returns a tensor of shape (n, Ny, Nx) where n is the number of Monte Carlo samples.
     """
-    mc_tensor = torch.zeros((n, sim_params.Ny, sim_params.Nx), dtype=torch.complex64, device=sim_params.device)
+    mc_tensor = torch.zeros((n, sim_params.Ny, sim_params.Nx), dtype=sim_params.dtype, device=sim_params.device)
 
     for i in range(n):
         u_init = u_init_func(sim_params, *u_init_func_args)
